Add a new module in order to manage modules.
This new module is able to register new modules that want to join the
koncentrator. Also add a few variables. One in this new module is used
in order to generate new ids for each incoming module. Another one is
located in app.py in order to list all available modules.