Update Makefile on rules stop, clean and deploy.
Fix all issues due to errors that were not errors. For example, the rule
stop could fail when one container was already stopped. Same when
removing an already removed container. This also fixes the case in which
the network cannot be created as it already exists causing make to
abort.