Update function get_content from class Module.
The order has been changed in order to handle the case in which the
content is expired. In this case, the koncentrator will do a request GET
on '/content' of the module.
This can prevent the content to be displayed by the view in the case
where the module has exited unexpectedly.