Update module 'hello-world-push'.
This update shows how to manage a module that could possibly have its
content outdated as it is still running. This is the reason why a Flask
server and a thread that runs the loop are started.
The server is started in order to handle requests in the case the
content has expired. The thread is used in order to regularly push
content on the koncentrator.