Add POST method on modules content.
This function is used for registered views in order to do a better cache
management and make faster for the views to update or not their
content. The view is still polling a specific module.
In the case of a push module, the Koncentrator is able to tell if the
current view has already polled it and if it needed that the view
updates its content.